Fixing warnings for a future compiler version, when the lab updates fedora version.
git-svn-id: svn+ssh://parasol-svn.cs.tamu.edu/research/parasol-svn/svnrepository/pmpl/trunk@4731 5e2c2758-810a-0410-8eb3-8e4e92d4004e

+++ src/Utilities/MedialAxisUtilities.h
@@ -382,7 +382,7 @@ ExactCollisionInfo(CfgType& _cfg, CfgType& _clrCfg,
     shared_ptr<ActiveMultiBody> robot = env->GetRobot(_cfg.GetRobotIndex());
 
     // Find closest point between robot and bbx, set if less than min dist from obstacles
-    for(int m=0; m < robot->NumFreeBody(); ++m) {
+    for(size_t m=0; m < robot->NumFreeBody(); ++m) {
       GMSPolyhedron& poly = robot->GetFreeBody(m)->GetWorldPolyhedron();
       for(size_t j = 0; j < poly.m_vertexList.size(); ++j){
         double clr = _bb->GetClearance(poly.m_vertexList[j]);
